pyproject.toml
[tool.poetry]
name = "agricite-backend"
version = "0.1.0"
description = ""
authors = ["Sunil Hanamshetty <64653515+shetty456@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
tortoise-orm = "^0.22.0"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
